# uint256: optimize Byte