--- a/uint256.go
+++ b/uint256.go
@@ -1236,20 +1236,18 @@ func (z *Int) Xor(x, y *Int) *Int {
 }
 
 // Byte sets z to the value of the byte at position n,
-// with 'z' considered as a big-endian 32-byte integer
-// if 'n' > 32, f is set to 0
-// Example: f = '5', n=31 => 5
+// with z considered as a big-endian 32-byte integer.
+// if n >= 32, z is set to 0
+// Example: z=5, n=31 => 5
 func (z *Int) Byte(n *Int) *Int {
 	// in z, z[0] is the least significant
 	//
-	if number, overflow := n.Uint64WithOverflow(); !overflow {
-		if number < 32 {
-			number := z[4-1-number/8]
-			offset := (n[0] & 0x7) << 3 // 8*(n.d % 8)
-			z[0] = (number & (0xff00000000000000 >> offset)) >> (56 - offset)
-			z[3], z[2], z[1] = 0, 0, 0
-			return z
-		}
+	if index, overflow := n.Uint64WithOverflow(); !overflow && index < 32 {
+		number := z[4-1-index/8]
+		offset := (index & 0x7) << 3 // 8 * (index % 8)
+		z[0] = (number >> (56 - offset)) & 0xff
+		z[3], z[2], z[1] = 0, 0, 0
+		return z
 	}
 	return z.Clear()
 }