The main difference is
```
z[0] = (number & (0xff00000000000000 >> offset)) >> (56 - offset)
```
vs
```
z[0] = (number >> (56 - offset)) & 0xff
```

------------------------
```
go test ./...
```
returns
```
ok  	github.com/holiman/uint256	0.972s
```
## Benchmark
code:
```
func BenchmarkByte(bench *testing.B) {
	a := new(Int).SetBytes(hex2Bytes("f123456789abcdeffedcba9876543210f2f3f4f5f6f7f8f9fff3f4f5f6f7f8f9"))
	n := NewInt(16)

	result := new(Int)
	bench.ResetTimer()
	for i := 0; i < bench.N; i++ {
		result.Set(a)
		result.Byte(n)
	}
}
```
```
goos: linux
goarch: amd64
pkg: github.com/holiman/uint256
cpu: AMD Ryzen 7 7735H with Radeon Graphics         
        │     old     │                new                 │
        │   sec/op    │   sec/op     vs base               │
Byte-16   2.849n ± 2%   2.571n ± 2%  -9.77% (p=0.000 n=10)

        │    old     │              new               │
        │    B/op    │    B/op     vs base            │
Byte-16   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal

        │    old     │              new               │
        │ allocs/op  │ allocs/op   vs base            │
Byte-16   0.000 ± 0%   0.000 ± 0%  ~ (p=1.000 n=10) ¹
¹ all samples are equal
```