@holiman 2024-05-06
```
name    old time/op    new time/op    delta
Byte-8     298ns ±42%     253ns ±29%  -15.07%  (p=0.049 n=10+10)
```
LGTM